Updating files
In the updated version, the Afriat Efficiency Index is calculated separately for each component of the set of observations, then the Afriat indices of the different components are aggregated using a maximum aggregator. It improves running time. Even though Afriat is quite fast anyway, we noticed that after improving the running time of Houtman Maks and of Varian, it became common that Afriat was the bottleneck, hence we did this improvement.