Update all Varian functions, now it will run much faster
Now we use intlinprog instead of my algorithm for the mean and meansqr aggregators. The algorithm for finding minimal cycles is intact though. 
Since now intlinprog takes no time, and the minimal cycles algorithm is the bottleneck, when calculating out-of-sample residuals we enumerate minimal cycles only once, thus saving a lot of running time.