Adapting to new Matlab version (R2019b)
Starting version R2019b Matlab does not allow to change the size of the GUI after it is initialized, so this function causes an error and a crash. I "fixed" it by adding an "if-else" that checks the version, and if it is 2019 or later it doesn't attempts to make the waitbar wider.